+++ demos/button.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Demonstrates how to use layouts.
+ * Demonstrates how to use buttons.
  */
 require 'init.php';
 use atk4\ui\Button;
@@ -80,3 +80,4 @@ $app->add(new Header(['Attaching', 'size' => 2]));
 $app->add(['Button', 'Scroll Up', 'top attached']);
 $app->add(['Table', 'attached', 'header' => false])->setSource(['One', 'Two', 'Three', 'Four']);
 $app->add(['Button', 'Scroll Up', 'bottom attached']);
+