+++ demos/checkbox.php
@@ -7,18 +7,18 @@ require '../vendor/autoload.php';
 try {
     $layout = new \atk4\ui\Layout\App(['defaultTemplate' => './templates/layout2.html']);
 
-    $layout->add(new \atk4\ui\Header(['Checkboxes', 'size' => 2]));
+    $layout->add(new \atk4\ui\Header(['CheckBoxes', 'size' => 2]));
 
-    $layout->add(new \atk4\ui\FormField\Checkbox('Make my profile visible'));
-
-    $layout->add(new \atk4\ui\View(['ui' => 'divider']));
-    $layout->add(new \atk4\ui\FormField\Checkbox(['Accept terms and conditions', 'slider']));
+    $layout->add(new \atk4\ui\FormField\CheckBox('Make my profile visible'));
 
     $layout->add(new \atk4\ui\View(['ui' => 'divider']));
-    $layout->add(new \atk4\ui\FormField\Checkbox(['Subscribe to weekly newsletter', 'toggle']));
+    $layout->add(new \atk4\ui\FormField\CheckBox(['Accept terms and conditions', 'slider']));
 
     $layout->add(new \atk4\ui\View(['ui' => 'divider']));
-    $layout->add(new \atk4\ui\FormField\Checkbox(['Custom setting?']))->js(true)->checkbox('set indeterminate');
+    $layout->add(new \atk4\ui\FormField\CheckBox(['Subscribe to weekly newsletter', 'toggle']));
+
+    $layout->add(new \atk4\ui\View(['ui' => 'divider']));
+    $layout->add(new \atk4\ui\FormField\CheckBox(['Custom setting?']))->js(true)->checkbox('set indeterminate');
 
     echo $layout->render();
 } catch (\atk4\core\Exception $e) {