+++ demos/database.php
@@ -73,7 +73,7 @@ if (!class_exists('Country')) {
             //$this->addField('description', ['ui'=>['form'=>['FormField/TextArea', 'rows'=>5]]]);
             $this->addField('description', ['type' => 'text']);
             $this->addField('client_name', ['type' => 'string']);
-            $this->addField('client_address', ['type' => 'string', 'ui' => ['form' => [new \atk4\ui\FormField\Textarea(), 'rows' => 4]]]);
+            $this->addField('client_address', ['type' => 'string', 'ui' => ['form' => [new \atk4\ui\FormField\TextArea(), 'rows' => 4]]]);
 
             $this->hasOne('client_country_iso', [
             new Country(),